@@ -1,290 +1,326 @@
-const chai = require('chai')
-const extract = require('../../src/extract-sub-schema')
-const expect = chai.expect
+import chai from 'chai';
+import extract from '../../src/extract-sub-schema';
+const expect = chai.expect;
 
-describe('extract sub schema', function() {
-  var schema
+describe.only('extract sub schema', function () {
+  var schema;
 
-  beforeEach(function() {
+  beforeEach(function () {
     schema = {
       properties: {
         name: {
-          minLength: 2
-        }
-      }
-    }
-  })
+          minLength: 2,
+        },
+      },
+    };
+  });
 
-  describe('plain schema', function() {
-    it('return a new schema the one sub schema', function() {
-      var result = extract(schema, '/properties/name')
-      expect(result).to.eql({minLength: 2})
-    })
+  describe('plain schema', function () {
+    it('return a new schema the one sub schema', function () {
+      var result = extract(schema, '/properties/name');
+      expect(result).to.eql({ minLength: 2 });
+    });
 
-    it('throws if not pointing to a subschema')
-  })
+    it('throws if not pointing to a subschema');
+  });
 
-  describe('non property pointer', function() {
-    it('extracts that schema', function() {
-      var result = extract({
-        properties: schema.properties,
-        additionalProperties: {
-          maxLength: 5,
-          allOf: [
-            {
-              pattern: '.+'
-            }
-          ]
-        }
-      }, '/additionalProperties')
+  describe('non property pointer', function () {
+    it('extracts that schema', function () {
+      var result = extract(
+        {
+          properties: schema.properties,
+          additionalProperties: {
+            maxLength: 5,
+            allOf: [
+              {
+                pattern: '.+',
+              },
+            ],
+          },
+        },
+        '/additionalProperties'
+      );
       expect(result).to.eql({
         maxLength: 5,
         allOf: [
           {
-            pattern: '.+'
-          }
-        ]
-      })
-    })
-  })
-
-  describe('not keyword', function() {
-    it('extracts the not schema', function() {
-      var result = extract({
-        properties: schema.properties,
-        not: {
-          properties: {
-            name: {
-              maxLength: 5
-            }
-          }
-        }
-      }, '/properties/name')
-      expect(result).to.eql({
-        minLength: 2,
-        not: {
-          maxLength: 5
-        }
-      })
-    })
-  })
-
-  describe('patternProperties', function() {
-    it('includes the schema if allOf if matching the property name', function() {
-      var result = extract({
-        properties: schema.properties,
-        patternProperties: {
-          'na.e': {
-            maxLength: 5
+            pattern: '.+',
           },
-          'n..e': {
-            pattern: '.+'
-          },
-          notMe: false
-        }
-      }, '/properties/name')
+        ],
+      });
+    });
+  });
 
-      expect(result).to.eql({
-        minLength: 2,
-        allOf: [
-          {
-            maxLength: 5
-          }, {
-            pattern: '.+'
-          }
-        ]
-      })
-    })
-  })
-
-  describe('additionalProperties', function() {
-    it('does not include the schema if matching properties or patternProperties', function() {
-      var result = extract({
-        properties: schema.properties,
-        patternProperties: {
-          'na.e': {
-            maxLength: 5
-          },
-          notMe: false
-        },
-        additionalProperties: {
-          pattern: '.+'
-        }
-      }, '/properties/name')
-
-      expect(result).to.eql({
-        minLength: 2,
-        allOf: [
-          {
-            maxLength: 5
-          }
-        ]
-      })
-    })
-
-    it('does include the schema if not matching properties or patternProperties', function() {
-      var result = extract({
-        properties: schema.properties,
-        patternProperties: {
-          'na.e': {
-            maxLength: 5
-          },
-          notMe: false
-        },
-        additionalProperties: {
-          pattern: '.+'
-        }
-      }, '/properties/foo')
-
-      expect(result).to.eql({
-        allOf: [
-          {
-            pattern: '.+'
-          }
-        ]
-      })
-    })
-
-    it('does include the schema if not matching properties or patternProperties', function() {
-      var result = extract({
-        properties: schema.properties,
-        patternProperties: {
-          'na.e': {
-            maxLength: 5
-          },
-          notMe: false
-        },
-        additionalProperties: false
-      }, '/properties/foo')
-
-      expect(result).to.eql({allOf: [false]})
-    })
-  })
-
-  describe('dependencies', function() {
-    it('considers ', function() {
-      var result = extract({
-        properties: schema.properties,
-        dependencies: {
-          other: {
+  describe('not keyword', function () {
+    it('extracts the not schema', function () {
+      var result = extract(
+        {
+          properties: schema.properties,
+          not: {
             properties: {
               name: {
-                maxLength: 7
-              }
-            }
-          }
-        }
-      }, '/properties/name', {presentKeys: ['other']})
-
-      expect(result).to.eql({
-        minLength: 2,
-        allOf: [
-          {
-            maxLength: 7
-          }
-        ]
-      })
-    })
-
-    it('works with nested dependencies')
-  })
-
-  describe('schemas in arrays', function() {
-    it('extracts allOf', function() {
-      var result = extract({
-        properties: schema.properties,
-        allOf: [
-          {
-            properties: {
-              name: {
-                maxLength: 5
-              }
-            }
-          }
-        ]
-      }, '/properties/name')
-      expect(result).to.eql({
-        minLength: 2,
-        allOf: [
-          {
-            maxLength: 5
-          }
-        ]
-      })
-    })
-
-    it('extracts anyOf', function() {
-      var result = extract({
-        properties: schema.properties,
-        anyOf: [
-          {
-            properties: {
-              name: {
-                maxLength: 5
-              }
-            }
-          }
-        ]
-      }, '/properties/name')
-      expect(result).to.eql({
-        minLength: 2,
-        anyOf: [
-          {
-            maxLength: 5
-          }
-        ]
-      })
-    })
-
-    it('extracts oneOf', function() {
-      var result = extract({
-        properties: schema.properties,
-        oneOf: [
-          {
-            properties: {
-              name: {
-                maxLength: 5
-              }
-            }
-          }
-        ]
-      }, '/properties/name')
-      expect(result).to.eql({
-        minLength: 2,
-        oneOf: [
-          {
-            maxLength: 5
-          }
-        ]
-      })
-    })
-
-    it('extracts deeply nested ones', function() {
-      var result = extract({
-        properties: schema.properties,
-        allOf: [
-          {
-            properties: {
-              name: {
-                maxLength: 5
-              }
+                maxLength: 5,
+              },
             },
-            allOf: [
-              {
-                properties: {
-                  name: {
-                    pattern: '.+'
-                  }
-                }
-              }, {
-                properties: {
-                  foo: true
-                }
-              }
-            ]
-          }
-        ]
-      }, '/properties/name')
+          },
+        },
+        '/properties/name'
+      );
+      expect(result).to.eql({
+        minLength: 2,
+        not: {
+          maxLength: 5,
+        },
+      });
+    });
+  });
+
+  describe('patternProperties', function () {
+    it('includes the schema if allOf if matching the property name', function () {
+      var result = extract(
+        {
+          properties: schema.properties,
+          patternProperties: {
+            'na.e': {
+              maxLength: 5,
+            },
+            'n..e': {
+              pattern: '.+',
+            },
+            notMe: false,
+          },
+        },
+        '/properties/name'
+      );
+
+      expect(result).to.eql({
+        minLength: 2,
+        allOf: [
+          {
+            maxLength: 5,
+          },
+          {
+            pattern: '.+',
+          },
+        ],
+      });
+    });
+  });
+
+  describe('additionalProperties', function () {
+    it('does not include the schema if matching properties or patternProperties', function () {
+      var result = extract(
+        {
+          properties: schema.properties,
+          patternProperties: {
+            'na.e': {
+              maxLength: 5,
+            },
+            notMe: false,
+          },
+          additionalProperties: {
+            pattern: '.+',
+          },
+        },
+        '/properties/name'
+      );
+
+      expect(result).to.eql({
+        minLength: 2,
+        allOf: [
+          {
+            maxLength: 5,
+          },
+        ],
+      });
+    });
+
+    it('does include the schema if not matching properties or patternProperties', function () {
+      var result = extract(
+        {
+          properties: schema.properties,
+          patternProperties: {
+            'na.e': {
+              maxLength: 5,
+            },
+            notMe: false,
+          },
+          additionalProperties: {
+            pattern: '.+',
+          },
+        },
+        '/properties/foo'
+      );
+
+      expect(result).to.eql({
+        allOf: [
+          {
+            pattern: '.+',
+          },
+        ],
+      });
+    });
+
+    it('does include the schema if not matching properties or patternProperties', function () {
+      var result = extract(
+        {
+          properties: schema.properties,
+          patternProperties: {
+            'na.e': {
+              maxLength: 5,
+            },
+            notMe: false,
+          },
+          additionalProperties: false,
+        },
+        '/properties/foo'
+      );
+
+      expect(result).to.eql({ allOf: [false] });
+    });
+  });
+
+  describe('dependencies', function () {
+    it('considers ', function () {
+      var result = extract(
+        {
+          properties: schema.properties,
+          dependencies: {
+            other: {
+              properties: {
+                name: {
+                  maxLength: 7,
+                },
+              },
+            },
+          },
+        },
+        '/properties/name',
+        { presentKeys: ['other'] }
+      );
+
+      expect(result).to.eql({
+        minLength: 2,
+        allOf: [
+          {
+            maxLength: 7,
+          },
+        ],
+      });
+    });
+
+    it('works with nested dependencies');
+  });
+
+  describe('schemas in arrays', function () {
+    it('extracts allOf', function () {
+      var result = extract(
+        {
+          properties: schema.properties,
+          allOf: [
+            {
+              properties: {
+                name: {
+                  maxLength: 5,
+                },
+              },
+            },
+          ],
+        },
+        '/properties/name'
+      );
+      expect(result).to.eql({
+        minLength: 2,
+        allOf: [
+          {
+            maxLength: 5,
+          },
+        ],
+      });
+    });
+
+    it('extracts anyOf', function () {
+      var result = extract(
+        {
+          properties: schema.properties,
+          anyOf: [
+            {
+              properties: {
+                name: {
+                  maxLength: 5,
+                },
+              },
+            },
+          ],
+        },
+        '/properties/name'
+      );
+      expect(result).to.eql({
+        minLength: 2,
+        anyOf: [
+          {
+            maxLength: 5,
+          },
+        ],
+      });
+    });
+
+    it('extracts oneOf', function () {
+      var result = extract(
+        {
+          properties: schema.properties,
+          oneOf: [
+            {
+              properties: {
+                name: {
+                  maxLength: 5,
+                },
+              },
+            },
+          ],
+        },
+        '/properties/name'
+      );
+      expect(result).to.eql({
+        minLength: 2,
+        oneOf: [
+          {
+            maxLength: 5,
+          },
+        ],
+      });
+    });
+
+    it('extracts deeply nested ones', function () {
+      var result = extract(
+        {
+          properties: schema.properties,
+          allOf: [
+            {
+              properties: {
+                name: {
+                  maxLength: 5,
+                },
+              },
+              allOf: [
+                {
+                  properties: {
+                    name: {
+                      pattern: '.+',
+                    },
+                  },
+                },
+                {
+                  properties: {
+                    foo: true,
+                  },
+                },
+              ],
+            },
+          ],
+        },
+        '/properties/name'
+      );
 
       expect(result).to.eql({
         minLength: 2,
@@ -293,12 +329,12 @@ describe('extract sub schema', function() {
             maxLength: 5,
             allOf: [
               {
-                pattern: '.+'
-              }
-            ]
-          }
-        ]
-      })
-    })
-  })
-})
+                pattern: '.+',
+              },
+            ],
+          },
+        ],
+      });
+    });
+  });
+});
